Changed the return type of DropHTTP receiveMessages to Collection<byte[]>
DropHTTP was sending out byte arrays but parsed received messages as
strings. This was modifiying the messages which results in the invalid
signature.

The DropTest is also modified. Class variables has been replaced with
methods variable, to avoid unintended behaviour when different tests are
using the same variables.

Also retrieveTest() is not an own JUnit test anymore, because these are
called in undefined order. Instead, this function is now called from the
send tests.

Since the keys are generated random for each run, we only expect one
drop message do be decryptable. Older messages were created with
different keys.