Fix erroneous escalation of invalid binary drop messages
It is not unlikely that binary drop messages are received that have the wrong
size compared to expectation implied by the header version byte.
This may for instance happen, if someone maliciously uploads invalid messages or
simply due to broken clients.

This patch fixes the handling of this such unexpected sizes.
They are no longer escalated to a RuntimeException (as was a bug), but are now
logged and then skipped.
Therefore, a new exception class is introduced to distinguish this size
mismatches from other critical version mismatches.