Refactored Persistence initialization
Replaced the init() method with an abstract connect() method. This is
called from the Persistence contructor.

Replaced the password argument from the constructor with a static
setPassword method in the Persistence class. This is just a workaround
until we know how to hand over the password to the persistence class,
but it is cleaner than changing all constructors.