Refactored Module class
Module class now doesn't get a EventEmitter in the constructor. Instead,
it can receive the required EventEmitter from its ModuleManager by a
package private method. This disallows direct access to the EventEmitter
for inheritors from Module which usually should not be in the module
package.

The required on() and emit() methods are wrapped for inheritors.