Include DropMessage type in Event name
Until now, all received DropMessages were send to everyone registered to
EVENT_DROP_MESSAGE_RECEIVED and the recipient had to filter out if he
wants to handle this message.

Now the event name contains the name of the DropMessage type and thus
only recipients registered to that type will receive the DropMessage.