Replaced DropMessage payload from ModelObject to String
Instead of ModelObject, Strings are now used as the payload for a
DropMessage. dropPayload is the actual payload while dropPayloadType
describes its type.

The advantage of this solution is, that the qabel-core doesn't has to
know the actual ModelObject type. Instead the a raw String is handed to
the receiver of a DropMessage. This allows a more loose coupling of
sender/receivers to the qabel-core. This is a little bit more
inconvenient for the sender and receiver of a DropMessage, because
serialization and deserialization isn't done by the qabel-core anymore.